PHR IPV4: Separate HandleIpv4 into its own file
This allows us to graphgen just the struct without worrying about mod
imports and other things that aren't really related.
